requirements.txt
xarg-python >= 1.5.1
dnspython
psutil
ping3
